Refatoração: alteração de double para Bigdecimal e atualização dos testes

+++ src/main/java/com/bankSystem/model/Transacao.java
@@ -1,15 +1,16 @@
 package com.bankSystem.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Transacao {
 
     private final LocalDateTime dataHora;
     private final TipoTransacao tipo;
-    private final double valor;
+    private final BigDecimal valor;
     private final String descricao;
 
-    public Transacao(TipoTransacao tipo, double valor, String descricao) {
+    public Transacao(TipoTransacao tipo, BigDecimal valor, String descricao) {
         this.dataHora = LocalDateTime.now();
         this.tipo = tipo;
         this.valor = valor;
@@ -24,7 +25,7 @@ public class Transacao {
         return tipo;
     }
 
-    public double getValor() {
+    public BigDecimal getValor() {
         return valor;
     }
