+++ src/test/java/com/bankSystem/model/TransacaoTest.java
@@ -2,22 +2,24 @@ package com.bankSystem.model;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 public class TransacaoTest {
     @Test
     void deveCriarTransacaoCorretamente() {
-        Transacao transacao = new Transacao(TipoTransacao.DEPOSITO, 100.0, "Depósito inicial");
+        Transacao transacao = new Transacao(TipoTransacao.DEPOSITO, BigDecimal.valueOf(100.0), "Depósito inicial");
 
         assertEquals(TipoTransacao.DEPOSITO, transacao.getTipo());
-        assertEquals(100.0, transacao.getValor());
+        assertEquals(BigDecimal.valueOf(100.0), transacao.getValor());
         assertEquals("Depósito inicial", transacao.getDescricao());
         assertNotNull(transacao.getDataHora());
     }
 
     @Test
     void toStringDeveRetornarFormatoCorreto() {
-        Transacao transacao = new Transacao(TipoTransacao.SAQUE, 50.0, "Retirada");
+        Transacao transacao = new Transacao(TipoTransacao.SAQUE, BigDecimal.valueOf(50.0), "Retirada");
         String toString = transacao.toString();
 
         assertTrue(toString.contains("SAQUE"));
